requirements.txt
openai~=0.26.1
python-telegram-bot~=20.1